Refactor out StreamTxStats and Losses
